requirements.txt
alembic==1.13.3
Flask==3.0.3
Flask_Login==0.6.3
Flask_Migrate==4.0.4
flask_sqlalchemy==3.1.1
flask-session
SQLAlchemy==2.0.35
Werkzeug==3.0.4
Gunicorn
psycopg2-binary
python-dotenv
Flask-Babel
gspread
google-auth-oauthlib